update:studentのe2eを書いた
